+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 Note: the BigBuy API is not versioned and some endpoints are often added or removed without notice.
 The removal of methods that call deleted endpoints is not considered a breaking change.
 
+## Unreleased
+
+* Bump `api-session` requirement to 1.4.0
+
 ## 3.19.6 (2024/04/16)
 
 * Don’t crash if BigBuy returns a bad error payload with inverted code and error message (`{"code": "Bad request", "message": 400}`)